refactor: refactored eligibility folder and added new folder under common
