+++ packages/e2e-tests/cypress/integration/eligibility/full-appeal/appellant-selects-the-site.feature
