+++ packages/e2e-tests/cypress/integration/eligibility/full-appeal/appellant-selects-the-site/appellant-selects-the-site-steps.js
@@ -1,16 +1,16 @@
 import { Given, Then } from 'cypress-cucumber-preprocessor/steps';
-import { verifyPageHeading } from '../../../support/common/verify-page-heading';
-import { verifyPageTitle } from '../../../support/common/verify-page-title';
-import { selectSiteOption } from '../../../support/eligibility/appellant-selects-the-site/select-site-option';
-import { getBackLink, getErrorMessageSummary } from '../../../support/common-page-objects/common-po';
-import { verifyDeselectSiteOption } from '../../../support/eligibility/appellant-selects-the-site/verify-deselect-site-option';
+import { verifyPageHeading } from '../../../../support/common/verify-page-heading';
+import { verifyPageTitle } from '../../../../support/common/verify-page-title';
+import { selectSiteOption } from '../../../../support/eligibility/appellant-selects-the-site/select-site-option';
+import { getBackLink, getErrorMessageSummary } from '../../../../support/common-page-objects/common-po';
+import { verifyDeselectSiteOption } from '../../../../support/eligibility/appellant-selects-the-site/verify-deselect-site-option';
 import {
   getNoneOfTheseOption,
   getSelectSiteErrorMessage,
-} from '../../../support/eligibility/page-objects/appellant-selects-the-site-po';
-import { verifyErrorMessage } from '../../../support/common/verify-error-message';
-import { goToAppealsPage } from '../../../support/common/go-to-page/goToAppealsPage';
-import { getContinueButton } from '../../../support/householder-planning/appeals-service/page-objects/common-po';
+} from '../../../../support/eligibility/page-objects/appellant-selects-the-site-po';
+import { verifyErrorMessage } from '../../../../support/common/verify-error-message';
+import { goToAppealsPage } from '../../../../support/common/go-to-page/goToAppealsPage';
+import { getContinueButton } from '../../../../support/householder-planning/appeals-service/page-objects/common-po';
 const pageHeading = 'Was your planning application was about any of the following?';
 const url = 'before-you-start/any-of-following';
 const pageTitle = 'Is your appeal about any of the following? - Before you start - Appeal a householder planning decision - GOV.UK';
