+++ packages/e2e-tests/cypress/integration/eligibility/full-appeal/enforcement-notice/enforcement-notice-steps.js
@@ -1,15 +1,15 @@
 import { Given, When, Then } from 'cypress-cucumber-preprocessor/steps';
-import { goToAppealsPage } from '../../../support/common/go-to-page/goToAppealsPage';
-import { verifyPageHeading } from '../../../support/common/verify-page-heading';
-import { verifyPageTitle } from '../../../support/common/verify-page-title';
+import { goToAppealsPage } from '../../../../support/common/go-to-page/goToAppealsPage';
+import { verifyPageHeading } from '../../../../support/common/verify-page-heading';
+import { verifyPageTitle } from '../../../../support/common/verify-page-title';
 import {
   getEnforcementNoticeErrorMessage,
   getEnforcementNoticeNo, getEnforcementNoticeYes,
-} from '../../../support/eligibility/page-objects/enforcement-notice-po';
-import { getErrorMessageSummary } from '../../../support/common-page-objects/common-po';
-import { verifyErrorMessage } from '../../../support/common/verify-error-message';
-import { getBackLink } from '../../../support/common-page-objects/common-po';
-import { getContinueButton } from '../../../support/householder-planning/appeals-service/page-objects/common-po';
+} from '../../../../support/eligibility/page-objects/enforcement-notice-po';
+import { getErrorMessageSummary } from '../../../../support/common-page-objects/common-po';
+import { verifyErrorMessage } from '../../../../support/common/verify-error-message';
+import { getBackLink } from '../../../../support/common-page-objects/common-po';
+import { getContinueButton } from '../../../../support/householder-planning/appeals-service/page-objects/common-po';
 const pageHeading = 'Have you received an enforcement notice?';
 const pageTitle = 'Have you received an enforcement notice? - Before you start - Appeal a planning decision - GOV.UK';
 const url = `before-you-start/enforcement-notice`;
