+++ packages/e2e-tests/cypress/integration/eligibility/full-appeal/granted-or-refused-application.feature
