+++ packages/e2e-tests/cypress/integration/eligibility/full-appeal/planning-application-type.feature
