+++ packages/e2e-tests/cypress/integration/eligibility/full-appeal/shutter-page.feature
