+++ packages/e2e-tests/cypress/integration/eligibility/full-appeal/shutter-page/shutter-page-steps.js
@@ -1,8 +1,8 @@
 import {Given} from 'cypress-cucumber-preprocessor/steps';
-import { verifyPageTitle } from '../../../support/common/verify-page-title';
-import { verifyPageHeading } from '../../../support/common/verify-page-heading';
-import { getACPLink } from '../../../support/eligibility/page-objects/shutter-page-po';
-import { goToAppealsPage } from '../../../support/common/go-to-page/goToAppealsPage';
+import { verifyPageTitle } from '../../../../support/common/verify-page-title';
+import { verifyPageHeading } from '../../../../support/common/verify-page-heading';
+import { getACPLink } from '../../../../support/eligibility/page-objects/shutter-page-po';
+import { goToAppealsPage } from '../../../../support/common/go-to-page/goToAppealsPage';
 const url = 'before-you-start/use-a-different-service';
 const pageHeading = 'You need to use a different service';
 const pageTitle = 'You need to use a different service - Before you start - Appeal a planning decision - GOV.UK';
